fix(e-loop-ledger): S6 loop 상세 stale-state — key={loopId} 리마운트 강제
까심 QA 적출: fetchAll 성공 분기에 setNotFound(false) 호출이 없고, Next App
Router가 같은 /loops/[id] 템플릿 인스턴스를 소프트-nav 시 재사용해 목록→상세A(404)
→목록→상세B 흐름에서 A의 notFound=true(+goal/brief)가 B에 잔존.

fix: 상세 로직을 loop-detail-client.tsx로 분리하고 page.tsx에서
<LoopDetailClient key={loopId}>로 강제 리마운트(PR #16 AuditClient·PR #19
TenantsClient와 동일 해법 — 이 세션 3번째 재발). route param으로 fetch하는 모든
상세 page는 key-remount가 기본이어야 한다.

재현: /loops/{존재안함}(404) → 목록 → /loops/{정상B} 클릭 시, fix 전엔 B도
"찾을 수 없습니다" 고정 → fix 후 B 정상 렌더 확인(로컬 isolated docker 재현).
